requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1

# LangChain split packages (no monolithic langchain)
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.1.0

# Embeddings / vector store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=4.0.0

# Pydantic (FastAPI dependency)
pydantic>=2.5.0
